QGIS server refactoring
Wraps the server into QgsServer class and introduces
Python bindings for the server.

All functions are now (static) class methods and the main
server program contains the FCGI loop that calls QgsServer
handleRequest.

The bindins can be used to run the server from Python and to
write tests for the server.